tests: start local squad server to support testing
Start a local instance of SQUAD so that Squad-Client tests can run
against it.

Add fixtures file, which provides all necessary data to be tested
in squad-client.

The mechanism is pretty simple, before every call to `./manage.py test`,
a fresh squad instance is started up and run that file so that
all data is available thru the api