squad_client: use module name for the logger name
In general it is a good practice to use the module name for the logger
such that it can be controlled within the python logging configuration
without modifying other loggers.

Right now, without this change, you have to change the root logger
configuration, which could impact other libraries or scripts.

Signed-off-by: Justin Cook <justin.cook@linaro.org>