all: explain metadata aliases
Sometimes the metadata variable we expect is not what tuxmake provides
in the `build.json`. So we create aliases of those values when building the metadata.

Some values also can have valid null values. If they do, we should try
to get the value from the CI environment if possible.

Signed-off-by: Justin Cook <justin.cook@linaro.org>