submit_tuxbuild: set schema version
Since we do not pin the version of jsonschema, an upstream change has
caused a problem with the pipeline.

jsonschema recently changed their default schema version to something
newer that is not compatiable with what we have defined in squad-client.

This sets the schema version to what it should be so that we can
investigate changing to the newer schema at a later date.

Signed-off-by: Justin Cook <justin.cook@linaro.org>