all: fetch testrun metadata from squad
Treat the testrun metadata like a squadobject, and use the built-in
features to send and recieve it.

Originally this was treated like a dictionary, and so there are some
changes related to that.

Because the testrun metadata does not really use an id, we remove the
uuid when we encode it to json.

Signed-off-by: Justin Cook <justin.cook@linaro.org>