squad_client: tux: change the way we name build tests
Here is now naming will work now.

Single builds use their build name, the first kconfig name, and a
combined kconfig sha, all using data from the tuxsuite build.json.

Ex: arm-clang12-defconfig-<sha>

To create the combined kconfig sha, we download all of the fragments
from the download_url (in build.json), combine them into a single file,
and than take the sha of that file.

Build sets use their build name and combine it with the target_arch,
toolchain, and first kconfig name, all using data from the tuxsuite
build.json.

Ex: arm-gcc-8-allnoconfig

To accomplish this we move this code into it's own module so we can
test it as a unit and share it with other tuxsuite-related projects if
needed.

Additionally, we need to fix a test that sets an environment variable
because it was having a side effect on other tests.

Signed-off-by: Justin Cook <justin.cook@linaro.org>