commands: submit_tuxbuild: update build test naming
After some discussion among members of the LKFT group, we have decided
to update how we name our build tests and their metrics. We hope that
these rules are easier to understand and apply.

If build_name is defined in the build json, use that when creating the
test name.

If build_name is not defined, use the target_arch and toolchain instead.

If kconfig has more than one entry in the build json, create a hash using
all the entries except the base config. The base config is the first
entry in kconfig.

If kconfig only has one entry, use only the base kconfig instead.

A format with the build_name:

<build_name>-<base_config>[-<kconfig_hash>]

A format without the build_name:

<target_arch>-<toolchain>-<base_config>[-<kconfig_hash>]

To achieve these rules, we created several functions to the
submit_tuxbuild command that could be imported and tested as a unit
instead of only through integration testing.

This will make it easier to replace the current build test naming scheme
by making sure everything is tested and implemented correctly.

Signed-off-by: Justin Cook <justin.cook@linaro.org>