tests: test_submit_tuxbuild: update tests
Update our tests to use new test data that includes newer tuxsuite data
and better models what we expect our build json to look like.

While doing this, I cleaned up the tests in a few different ways.

Test the metadata of each testrun of a build instead of the entire
build's metadata at once. I think this makes it a little easier to
understand and is more inline with how the command works in additiona to
making it easier to understand what is being tested and why.

Remove some json schema tests. I think these tests do a better job of
testing the jsonschema library than our own code (which we're not
interested in doing), so I think we should remove them.

Remove deprecated test data that is no longer being used.

Signed-off-by: Justin Cook <justin.cook@linaro.org>