squad_client: commands: submit_tuxbuild: update build test naming
After some discussion among lkft members, we decided to update how we
name our build tests and their metrics.

If a build_name is defined in the build json, use that to create the
build test name.

If build_name is not defined, use the target_arch, toolchain, the first
kconfig and potentially a kconfig sha based on the names of the
remaining kconfig files and values.

Example:

<target_arch>-<toolchain>-<base_config>[-<kconfig_sha>]

x86_64-gcc-11-defconfig

If a kconfig sha is present, it will always be 8 characters in length.

Signed-off-by: Justin Cook <justin.cook@linaro.org>