core: models: fix initialization of class variables
Initialize class variables with None and then set them to a mutable type
when the instance calls a method.

If you initialize a class variable with a mutable type, it will be
shared amonst all instances of the class, rather than being instance
specific.

This means a change to that variable in one instance will affect all
other instances.

https://docs.python.org/3/tutorial/classes.html#class-and-instance-variables

"The concept of class-variable and instance-variable is fine till the
point where you use mutable objects".

Suggested-by: Charles Oliveira <charles.oliveira@linaro.org>

Signed-off-by: Justin Cook <justin.cook@linaro.org>