squad_client/core/models: Fix TestRun init function
The feature that allows a TestRun ID to be passed to the TestRun
constructor was accidentally removed during a recent update.

As a result, the TestRun init function was missing option to initialise
a TestRun object by an existing TestRun ID.

Add ID as an optional parameter to the TestRun init function to restore
this functionality.

Move the super().__init__(_id) so it is called after the
default values of the object's members have been initialised to None or
empty. This means that the data will be filled in if an ID is provided
and initialised to None or [] if no ID is provided.

Signed-off-by: Katie Worton <katie.worton@linaro.org>